Перенесена функция фильтрации объявлений в модуль filter.js
