+++ js/generation-similar.js
@@ -1,3 +1,5 @@
+import {filterOffers} from './filter.js';
+
 const typeHousing = {
   flat: 'Квартира',
   bungalow: 'Бунгало',
@@ -6,15 +8,7 @@ const typeHousing = {
   hotel: 'Отель',
 };
 
-const Price = {
-  LOW: 10000,
-  HIGH: 50000,
-};
-
-const DEFAULT_VALUE = 'any';
-
 const SIMILAR_DESCRIPTION_COUNT = 10;
-
 const cardTemplate = document.querySelector('#card').content.querySelector('.popup');
 
 const getType = (type) => typeHousing[type];
@@ -51,71 +45,6 @@ const checkParameter = (template, parameter, text, className) => {
   const element = template.querySelector(className);
 
   (parameter) ? element.textContent = text : hideElement(element);
-};
-
-const getCheckedFeatures = (array) => {
-  const checkedArray = [];
-
-  array.forEach((element) => {
-    if (element.checked) {
-      checkedArray.push(element.value);
-    }
-  });
-
-  return checkedArray;
-};
-
-const filterOffers = (property) => {
-  const housingTypeFilter = document.querySelector('select[name="housing-type"]');
-  const housingPriceFilter = document.querySelector('select[name="housing-price"]');
-  const housingRoomsFilter = document.querySelector('select[name="housing-rooms"]');
-  const housingGuestsFilter = document.querySelector('select[name="housing-guests"]');
-  const housingMapFeatures = document.querySelectorAll('.map__checkbox');
-
-  const checkedFeatures = getCheckedFeatures(housingMapFeatures);
-
-  let isValidType = false;
-  let isValidPrice = false;
-  let isValidRooms = false;
-  let isValidGuests = false;
-  let isValidFeatures = true;
-
-  isValidType = (property.offer.type === housingTypeFilter.value || housingTypeFilter.value === DEFAULT_VALUE);
-
-  switch (true) {
-    case housingPriceFilter.value === DEFAULT_VALUE:
-      isValidPrice = true;
-      break;
-    case housingPriceFilter.value === 'middle' && property.offer.price >= Price.LOW && property.offer.price < Price.HIGH:
-      isValidPrice = true;
-      break;
-    case housingPriceFilter.value === 'low' && property.offer.price < Price.LOW:
-      isValidPrice = true;
-      break;
-    case housingPriceFilter.value === 'high' && property.offer.price >= Price.HIGH:
-      isValidPrice = true;
-      break;
-  }
-
-  if (property.offer.rooms === Number(housingRoomsFilter.value) || housingRoomsFilter.value === DEFAULT_VALUE) {
-    isValidRooms = true;
-  }
-
-  if (property.offer.guests === Number(housingGuestsFilter.value) || housingGuestsFilter.value === DEFAULT_VALUE) {
-    isValidGuests = true;
-  }
-
-  if (checkedFeatures.length) {
-    if (property.offer.features) {
-      isValidFeatures = checkedFeatures.every((feature) => property.offer.features.includes(feature));
-    } else {
-      isValidFeatures = false;
-    }
-  }
-
-  if (isValidType && isValidPrice && isValidRooms && isValidGuests && isValidFeatures) {
-    return property;
-  }
 };
 
 const generateSimilarProperties = (descriptionOffer) => {
